project/apps/items-validator/requirements.txt
faust-streaming
confluent-kafka
confluent-kafka[json]
orjson